+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - added file watch mode setting to configure file change detection method
 - properly focusing search input in task sidebar
 - added direct support for Neuralwatt provider
+- added getProviders to extension context to allow fetching the configured providers
 
 ## [0.66.0]